Adding packaged filed
