+++ params.py
@@ -1,0 +1,5 @@
+path_to_exec = '/Users/marida/Downloads/chromedriver'
+account = 'blueapron' #IG account name
+csv_name = f'followers_{account}.csv'
+usr = ''
+pwd = ''
